Move details editor and logs component to shared location
