+++ src/components/tables/Details/EditorAndLogs.tsx
